scripts/requirements.txt
async-timeout==4.0.2
Deprecated==1.2.13
numpy==1.22.3
packaging==21.3
pandas==1.4.2
psycopg2==2.9.3
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
six==1.16.0
wrapt==1.14.0
